+++ src/application/Randomizer.cpp
@@ -71,6 +71,17 @@ std::map<int, LocationData> Randomizer::GetCheckData() const
     return _locationRepository.GetLocations();
 }
 
+std::vector<LifeBoxLocationData> Randomizer::GetLifeCapsules()
+{
+    
+    return _locationRepository.GetLifeCapsules();
+}
+
+void Randomizer::OnLifeSanitySet(bool lifeSanity)
+{
+    _displayManager.SetLifeSanity(lifeSanity);
+}
+
 void Randomizer::OnConnected()
 {
     _displayManager.QueueMessage("Connected to Archipelago");
@@ -103,3 +114,4 @@ void Randomizer::SetMissions(Characters characters, int missions)
 {
     _displayManager.SetMissions(characters, missions);
 }
+