+++ src/application/Randomizer.h
@@ -35,6 +35,8 @@ public:
     void QueueNewMessage(std::string information);
     void OnEmblemGoalSet(int emblemGoal);
     void SetMissions(Characters characters, int missions);
+    std::vector<LifeBoxLocationData> GetLifeCapsules();
+    void OnLifeSanitySet(bool lifeSanity);
 
 private:
     DisplayManager& _displayManager;